% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BaseClasses.R
\docType{class}
\name{HW2Config-class}
\alias{HW2Config}
\alias{HW2Config-class}
\alias{populate,HW2Config-method}
\alias{subjectName,HW2Config-method}
\title{HitWalker2 Configuration}
\usage{
\S4method{subjectName}{HW2Config}(obj)

\S4method{populate}{HW2Config}(obj, neo.path = NULL, skip = NULL)
}
\description{
A Class Representing a HitWalker2 Configuration
}
\section{Methods (by generic)}{
\itemize{
\item \code{subjectName}: Extracts the subject names defined in a HW2Config object

\item \code{populate}: Populates a neo4j database.  The subject/sample info is populated first followed by the remaining entries.
If \code{neo.path} is specified, the neo4j-shell executable is expected at neo.path/bin/neo4j-shell.  Otherwise it is expected to be part of
your path.  If \code{skip} is specified, it should be an integer vector indicating which of entries in the \code{data.list} slot should be skipped.
}}
\section{Slots}{

\describe{
\item{\code{subject}}{A Subject object to be used as the basis for this HitWalker2 database}

\item{\code{data.list}}{Named list containing the experiemntal data}

\item{\code{data.types}}{A list of the form: list(seeds=seed.vec,target='target') where the names in seed and target correspond to the names in data.list}

\item{\code{gene.model}}{String naming the type of gene model to be used, currently only entrez is supported.}
}}

