% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BaseClasses.R
\docType{class}
\name{HW2Config-class}
\alias{HW2Config}
\alias{HW2Config-class}
\alias{configure,HW2Config-method}
\alias{populate,HW2Config-method}
\alias{subjectName,HW2Config-method}
\title{HitWalker2 Configuration}
\usage{
\S4method{subjectName}{HW2Config}(obj)

\S4method{populate}{HW2Config}(obj, neo.path = NULL, skip = NULL)

\S4method{configure}{HW2Config}(obj,
  base.dir = "/home/vagrant/HitWalker2/populate/",
  dest.dir = "/home/vagrant/HitWalker2/network/", make.graph.struct = T)
}
\arguments{
\item{obj}{An object of class \code{HW2Config}}

\item{skip}{If \code{neo.path} is specified, the neo4j-shell executable is expected at neo.path/bin/neo4j-shell.  Otherwise it is expected to be part of your path.}

\item{base.dir}{Directory where the template files are located}

\item{dest.dir}{Directory where the complete template files should be placed.}

\item{make.graph.struct}{If TRUE, generate and stage a graph_struct file}

\item{skip}{If \code{skip} is specified, it should be an integer vector indicating which of entries in the \code{data.list} slot should be skipped.}
}
\description{
A class representing a HitWalker2 configuration
}
\section{Methods (by generic)}{
\itemize{
\item \code{subjectName}: Extracts the subject names defined in a HW2Config object

\item \code{populate}: Populates a neo4j database.  The subject/sample info is populated first followed by the remaining entries.

\item \code{configure}: Configures a HitWalker2 instance by substituting values into the template files defined in 'base.dir' and placing them into 'dest.dir'.
The defaults should suffice for Vagrant-based HitWalker2 instances.
}}
\section{Slots}{

\describe{
\item{\code{subject}}{A Subject object to be used as the basis for this HitWalker2 database}

\item{\code{data.list}}{Named list containing the experiemntal data}

\item{\code{data.types}}{A list of the form: list(seeds=seed.vec,target='target') where the names in seed and target correspond to the names in data.list}

\item{\code{gene.model}}{String naming the type of gene model to be used, currently only entrez is supported.}
}}

